Refactoring, SBT implementation
Refactoring, SBT implementation

+++ contracts/IWRLD_Records.sol
@@ -5,22 +5,22 @@ interface IWRLD_Records {
   struct StringRecord {
     string value;
     string typeOf;
-    uint256 ttl;
+    uint32 ttl;
   }
 
   struct AddressRecord {
     address value;
-    uint256 ttl;
+    uint32 ttl;
   }
 
   struct UintRecord {
     uint256 value;
-    uint256 ttl;
+    uint32 ttl;
   }
 
   struct IntRecord {
     int256 value;
-    uint256 ttl;
+    uint32 ttl;
   }
 
 }
